[shared] Change name

+++ apps/statistics/store.cpp
@@ -442,12 +442,12 @@ double Store::computeModes(int series, int i, double * modeFreq, int * modesTota
   return ithValue;
 }
 
-bool Store::deleteValueAtIndex(int series, int i, int j, bool safeDeletion, bool delayUpdate) {
-  if (safeDeletion) {
-    return DoublePairStore::deleteValueAtIndex(series, i, j, safeDeletion, delayUpdate);
+bool Store::deleteValueAtIndex(int series, int i, int j, bool authorizeNonEmptyRowDeletion, bool delayUpdate) {
+  if (authorizeNonEmptyRowDeletion) {
+    deletePairOfSeriesAtIndex(series, j, delayUpdate);
+    return true;
   }
-  deletePairOfSeriesAtIndex(series, j, delayUpdate);
-  return true;
+  return DoublePairStore::deleteValueAtIndex(series, i, j, authorizeNonEmptyRowDeletion, delayUpdate);
 }
 
 /* Private methods */
